% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alert_range_notify.R
\name{alert_range_notify}
\alias{alert_range_notify}
\title{This function reject all values greater than max and smaller than min in a data.frame}
\usage{
alert_range_notify(DATA, DATETIME_HEADER = "TIMESTAMP",
  DATETIME_FORMAT = "\%Y-\%m-\%d \%H:\%M", RECORD_HEADER, RANGE_DIR,
  RANGE_FILE, MAIL_DIR, MAIL_FILE_ALERT, STATION, USE_FLAG, USE_RT_FLAG)
}
\arguments{
\item{DATA}{data.frame of timeseries}

\item{DATETIME_HEADER}{header of data corresponding to datetime}

\item{DATETIME_FORMAT}{kkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkkk}

\item{RANGE_DIR}{directory where support files are stored}

\item{RANGE_FILE}{name of filr where min/max thresholds are defined for each variable. Thi file is in RANGE_DIR}

\item{MAIL_DIR}{aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa}

\item{MAIL_FILE_ALERT}{aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa}

\item{STATION}{name of station}

\item{USE_FLAG}{true/false --> decide to use or not the station flags in range file}

\item{USE_RT_FLAG}{true/false --> decide to use or not the station flags in range file}
}
\value{
A data.frame with all values in its physical range
}
\description{
This function reject all values greater than max and smaller than min in a data.frame
}
\examples{
exclude_out_of_range(DATA = mydata, RANGE_DIR = RANGE_DIR, RANGE_FILE = "Range.csv")

}
