% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_empty_files.R
\name{check_empty_files}
\alias{check_empty_files}
\title{This function check the file size of input file}
\usage{
check_empty_files(SCHEDULING_DIR, FILE)
}
\arguments{
\item{SCHEDULING_DIR}{Directory where input file is stored}

\item{FILE}{File name (.dat)}
}
\description{
This function check the file size of input file
}
\examples{
check_empty_files(SCHEDULING_DIR = "~/Data/Input/",FILE = "M4s.dat")
check_empty_files(SCHEDULING_DIR = "Your input folder",FILE = "Your file to check")
}
