% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pics_corrupt.R
\name{corr_pics}
\alias{corr_pics}
\title{This function check pictures corruption. If the lower half of the picture is on the same color (= identical value) the picture is classified corrupted}
\usage{
corr_pics(pics)
}
\arguments{
\item{pics}{The extended path and file name of the picture}
}
\value{
A logical value that indicate if the pic is corrupted or not (corrupted == TRUE)
}
\description{
This function check pictures corruption. If the lower half of the picture is on the same color (= identical value) the picture is classified corrupted
}
