% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletes_duplcated_data.R
\name{deletes_duplcated_data}
\alias{deletes_duplcated_data}
\title{This function delete duplcated data in a data.frame}
\usage{
deletes_duplcated_data(DATA)
}
\arguments{
\item{DATA}{data.frame}
}
\value{
A data.frame without duplicated rows
}
\description{
This function delete duplcated data in a data.frame
}
\examples{
deletes_duplcated_data(DATA = mydata)
deletes_duplcated_data(DATA = your data.framel)

}
