% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletes_old_datetime.R
\name{deletes_old_datetime}
\alias{deletes_old_datetime}
\title{This function delete duplcated data in a data.frame}
\usage{
deletes_old_datetime(DATA, DATETIME_HEADER = "TIMESTAMP")
}
\arguments{
\item{DATA}{data.frame}

\item{DATETIME_HEADER}{header corresponding to datetime}
}
\value{
A list containing a data.frame without dates already analyzed rows and a dataframe with dates already analyzed
}
\description{
This function delete duplcated data in a data.frame
}
\examples{
deletes_old_datetime(DATA = mydata, DATETIME_HEADER = "TIMESTAMP")
deletes_old_datetime(DATA = your data.frame, DATETIME_HEADER = "Your datetime header")

}
