% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_overlap.R
\name{detect_overlap}
\alias{detect_overlap}
\title{This function detect overlapping dates in a data.frame}
\usage{
detect_overlap(DATA, DATETIME_HEADER = "TIMESTAMP",
  RECORD_HEADER = "RECORD")
}
\arguments{
\item{DATA}{data.frame}

\item{DATETIME_HEADER}{header corresponding to datetime}

\item{RECORD_HEADER}{header corresponding to record}
}
\value{
A data.frame containing records and datetime where overlaps occour
}
\description{
This function detect overlapping dates in a data.frame
}
\examples{
detect_overlap(mydata, "TIMESTAMP", "RECORD")
detect_overlap(DATA = mydata, DATETIME_HEADER = "your datetime header", RECORD_HEADER = "your record header")

}
