% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DQC_function.R
\name{DQC_function}
\alias{DQC_function}
\title{This function checks a data file evaluating structures problems, missing and overlapping data, overrange data and writing different data files for different pourpuse. 
It is the core of the DQC scripts. It can be applied for every single files or in a operative method for a near real time monitoring. 
This function use different functions and process to manage data append and structured storage.}
\usage{
DQC_function(input_dir, output_dir_data, output_dir_report, project_dir,
  data_from_row, header_row_number, datetime_header, datetime_format,
  datetime_sampling, record_header, range_file, write_output_files,
  write_output_report, file_name, station_name, start_date,
  logger_info_file, record_check, output_dir_raw, use_alert_station_flag,
  mail_file_alert, use_realtime_station_flag, header_check, sep)
}
\arguments{
\item{input_dir}{character: is the directory where to source input data file}

\item{output_dir_data}{character: is the directory where total data are stored}

\item{output_dir_report}{NOT USED! character:  It is the directory where report/alert files are stored}

\item{project_dir}{NOT USED! character: It is the main directory used to distinguish different project}

\item{data_from_row}{numeric: when you open a file it is the numer of the first row of data (header excluded!)}

\item{header_row_number}{numeric: when you open a file it iis the numeber of the row indicating the real header or column names}

\item{datetime_header}{character: is the name of the column indicates the datetime. For examples "TIMESTAMP"}

\item{datetime_format}{character: is the datetime format. For examples "%Y-%m-%d %H:%M"}

\item{datetime_sampling}{character: is the time sampling interval. For examples "15 min"}

\item{record_header}{character: is the name of the column indicates the record. For examples "RECORD"}

\item{range_file}{character: is the name of the file containg the limits of overrange data  # NOTE --> add range_dir to input of DQC_function and adapt all scripts!}

\item{write_output_files}{logical: is a flag that indicates the necessity of writing the total/raw files.}

\item{write_output_report}{NOT USED! logical: is a flag that indicates the necessity of writing the report files.}

\item{file_name}{character: is the file name of file to process}

\item{station_name}{character: is the name of station to process. It could be extracted with rules from file_name}

\item{start_date}{POSIXct: is the date of the first data to check}

\item{logger_info_file}{character: is the name of the file used to compare headers and higlight changes}

\item{record_check}{numeric: is a flag that indicates if check (record_check = 1) or not (record_check = 0) if the records are continuous}

\item{output_dir_raw}{character: is the directory where raw data are stored}

\item{use_alert_station_flag}{logical: is a flag to decide to use or not the station flags in range file # --> To clarify the usage}

\item{mail_file_alert}{character: is the name of the file that indicates if to check overrange and  to manage sending mail}

\item{header_check}{logical: is a flag that indicates if you wanto to check header and logger information (based on new station file standard) (TRUE --> check all!)}

\item{sep}{character: is the character indicating the separator in the input file}

\item{use_realtime_station_flag:}{is a flag indicates to decide to use or not the station flags in range file # --> To clarify the usage}
}
\value{
A list. The list contains: 1. A dataframe of th complete header 2. A dataframe of data column names 3. A dataframe of data 4. A list of errors 5.A dataframe of data without out of range data
}
\description{
This function checks a data file evaluating structures problems, missing and overlapping data, overrange data and writing different data files for different pourpuse. 
It is the core of the DQC scripts. It can be applied for every single files or in a operative method for a near real time monitoring. 
This function use different functions and process to manage data append and structured storage.
}
