% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_in_scheduling_dir.R
\name{files_in_scheduling_dir}
\alias{files_in_scheduling_dir}
\title{This function explore the file in a specific folder}
\usage{
files_in_scheduling_dir(SCHEDULING_DIR = scheduling_dir)
}
\arguments{
\item{SCHEDULING_DIR}{directory where files are stored}
}
\value{
a vector of files available in SCHEDULING_DIR
}
\description{
This function explore the file in a specific folder
}
\examples{
files_in_scheduling_dir(SCHEDULING_DIR = scheduling_dir)
files_in_scheduling_dir(SCHEDULING_DIR = scheduling_dir)

}
