% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_sensor_KO.R
\name{find_sensor_KO}
\alias{find_sensor_KO}
\title{This function reject all values greater than max and smaller than min in a data.frame}
\usage{
find_sensor_KO(DATA, DATETIME_HEADER = "TIMESTAMP", SUPPORT_DIR, RANGE_FILE)
}
\arguments{
\item{DATA}{data.frame of timeseries}

\item{DATETIME_HEADER}{header of data corresponding to datetime}

\item{SUPPORT_DIR}{directory where support files are stored}

\item{RANGE_FILE}{name of filr where min/max thresholds are defined for each variable. Thi file is in SUPPORT_DIR}
}
\value{
A data.frame with all values in its physical range
}
\description{
This function reject all values greater than max and smaller than min in a data.frame
}
\examples{
exclude_out_of_range(DATA = mydata, SUPPORT_DIR = support_dir, RANGE_FILE = "Range.csv")

}
