% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hobo_to_campbell.R
\name{hobo_to_campbell}
\alias{hobo_to_campbell}
\title{This function convert HOBO header in a the CambpellScientific TOA5 format}
\usage{
hobo_to_campbell(HEADER, DATETIME_HEADER, RECORD_HEADER)
}
\arguments{
\item{HEADER}{is the header of input file}

\item{DATETIME_HEADER}{is the name of the column indicates datetime}

\item{RECORD_HEADE}{is the name of the column indicates record}
}
\value{
Return the new header in the CampbellScientific TOA5 format
}
\description{
This function convert HOBO header in a the CambpellScientific TOA5 format
}
