% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_dates.R
\name{missing_dates}
\alias{missing_dates}
\title{This function detect missing dates in a data.frame timeserie and fill with rows containing NAs}
\usage{
missing_dates(DATA, DATETIME_HEADER = DATETIME_HEADER,
  RECORD_HEADER = RECORD_HEADER, DATETIME_SAMPLING = DATETIME_SAMPLING)
}
\description{
@param DATA data.frame having a column defined as POSIXct (datetime)
 @param DATETIME_HEADER header corresponding to datetime
 @param RECORD_HEADER header corresponding to record
 @param DATETIME_SAMPLING time sampling (e.g. "15 min", "hour". See seq.POSIXt {base} on website)
 
 @return a data.frame which contains a continuos timeseries
 
 @export
 
 @examples
 missing_dates(DATA = mydata ,DATETIME_HEADER = "TIMESTAMP", RECORD_HEADER = "Record", DATETIME_SAMPLING = "15 min")
 missing_dates(DATA = your data.frame ,DATETIME_HEADER = "your datetime header", RECORD_HEADER = "your datetime record", DATETIME_SAMPLING = "your datetime sampling")
}
