% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_record.R
\name{missing_record}
\alias{missing_record}
\title{This function detect missing record in a data.frame timeserie and fill with rows containing NAs}
\usage{
missing_record(DATA, DATETIME_HEADER = DATETIME_HEADER,
  RECORD_HEADER = RECORD_HEADER, DATETIME_SAMPLING = DATETIME_SAMPLING,
  DATETIME_FORMAT = DATETIME_FORMAT)
}
\description{
@param DATA data.frame having a column defined as POSIXct (datetime)
 @param DATETIME_HEADER header corresponding to datetime
 @param RECORD_HEADER header corresponding to record
 @param DATETIME_SAMPLING time sampling (e.g. "15 min", "hour". See seq.POSIXt {base} on website)
 @param DATETIME_FORMAT datetime format ("%Y-%m-%d %h:%m")
 @return a data.frame which contains a continuos timeseries
}
\details{
@export

@examples
 missing_record(DATA = mydata ,DATETIME_HEADER = "TIMESTAMP", RECORD_HEADER = "Record", DATETIME_SAMPLING = "15 min")
 missing_record(DATA = your data.frame ,DATETIME_HEADER = "your datetime header", RECORD_HEADER = "your datetime record", DATETIME_SAMPLING = "your datetime sampling")
}
