% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_update_download_table.R
\name{read_and_update_download_table}
\alias{read_and_update_download_table}
\title{This function check the file size of input file}
\usage{
read_and_update_download_table(DOWNLOAD_TABLE_DIR, FILES_AVAILABLE)
}
\arguments{
\item{DOWNLOAD_TABLE_DIR}{Directory where download table is.}

\item{FILES_AVAILABLE}{File name (.dat) in a specific folder.}
}
\value{
A dataframe called download_table which contain the name of station, the last download data, the last data file modification and a flag used to avoid Data Quality Check
}
\description{
This function check the file size of input file
}
\examples{
check_empty_file(INPUT_DATA_DIR = "~/Data/Input/",FILE = "M4s.dat")
check_empty_file(INPUT_DATA_DIR = "Your input folder",FILE = "Your file to check")

}
