% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_update_issue_counter.R
\name{read_and_update_issue_counter}
\alias{read_and_update_issue_counter}
\title{This function check the file size of input file}
\usage{
read_and_update_issue_counter(ISSUE_COUNTER_DIR, FILES_AVAILABLE,
  DATETIME_FORMAT, PROJECT = NA)
}
\arguments{
\item{ISSUE_COUNTER_DIR}{Directory where download table is.}

\item{FILES_AVAILABLE}{File name (.dat) in a specific folder.}

\item{DATETIME_FORMAT}{Datetime format}

\item{PROJECT}{}
}
\value{
A dataframe called download_table which contain the name of station, the last download data, the last data file modification and a flag used to avoid Data Quality Check
}
\description{
This function check the file size of input file
}
\examples{
check_empty_file(INPUT_DATA_DIR = "~/Data/Input/",FILE = "M4s.dat", DATETIME_FORMAT = "\%Y-\%m-\%d \%H:\%M")
check_empty_file(INPUT_DATA_DIR = "Your input folder",FILE = "Your file to check")

}
