% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_download_table_pics.R
\name{read_download_table_pics}
\alias{read_download_table_pics}
\title{This function check the file size of input file}
\usage{
read_download_table_pics(DOWNLOAD_TABLE_DIR, DOWNLOAD_TABLE_FILE,
  FOLDERS_AVAILABLE, DATETIME_FORMAT, PROJECT = NA)
}
\arguments{
\item{DOWNLOAD_TABLE_DIR}{Directory where download table is.}

\item{FOLDERS_AVAILABLE}{File name (.dat) in a specific folder.}

\item{DATETIME_FORMAT}{Datetime format}

\item{PROJECT}{}
}
\value{
A dataframe called download_table which contain the name of station, the last download data, the last data file modification and a flag used to avoid Data Quality Check
}
\description{
This function check the file size of input file
}
\examples{
check_empty_file(INPUT_DATA_DIR = "~/Data/Input/",FILE = "M4s.dat", DATETIME_FORMAT = "\%Y-\%m-\%d \%H:\%M")
check_empty_file(INPUT_DATA_DIR = "Your input folder",FILE = "Your file to check")

}
