% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structure_control.R
\name{structure_control}
\alias{structure_control}
\title{This function check if in input file there is empty rows and if a row is shifted in the previous line}
\usage{
structure_control(FILE_PATH, FILE_NAME, DATETIME_HEADER = "TIMESTAMP",
  DATETIME_FORMAT = "yyyy-mm-dd HH:MM", DATA_FROM_ROW = 5,
  HEADER_ROW_NUMBER = 2)
}
\description{
@param FILE_PATH directory where input file is stored
 @param FILE_NAME name of file to read. Admitted files .csv and .dat
 @param DATETIME_HEADER header corresponding to datetime
 @param DATETIME_FORMAT format of datetime (E.g. "yyyy-mm-dd HH:MM")
 @param DATA_FROM_ROW the number of row of the first value
 @param HEADER_ROW_NUMBER the number of row of the header
}
\details{
@return a list containing a data.frame of header, a data.frame of column names, a data.frame of data

@export

@examples
 read_data(FILE_PATH = "~/Data/Input/", FILE_NAME = "M4s.dat", DATETIME_HEADER = "TIMESTAMP" , DATETIME_FORMAT = "yyyy-mm-dd HH:MM", DATA_FROM_ROW = 5, HEADER_ROW_NUMBER = 2)
 read_data(FILE_PATH = "Your input file storage", FILE_NAME = "Your data name", DATETIME_HEADER = "Your datetime headere" , DATETIME_FORMAT = "Your datetime format", DATA_FROM_ROW = "The row of your first data", HEADER_ROW_NUMBER = "The row of your data column names")
}
