% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_to_char.R
\name{time_to_char}
\alias{time_to_char}
\title{This function convert POSIXct datetime in a data.frame to character}
\usage{
time_to_char(DATA, DATETIME_HEADER, DATETIME_FORMAT)
}
\description{
@param DATA data.frame having a column defined as POSIXct (datetime)
 @param DATETIME_HEADER header corresponding to datetime
 @param DATETIME_FORMAT format of datetime (E.g. "yyyy-mm-dd HH:MM")
 
 @return a data.frame which contains character datetime
 
 @export
 
 @examples
 time_to_char = function(DATA = mydata, DATETIME_HEADER = "TIMESTAMP", DATETIME_FORMAT = "yyyy-mm-dd HH:MM")
 time_to_char = function(DATA = your data.frame, DATETIME_HEADER = "your datetime header", DATETIME_FORMAT = "your datetime format")
}
