% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deletes_duplcated_data.R
\name{deletes_duplcated_data}
\alias{deletes_duplcated_data}
\title{This function delete duplcated data in a data.frame}
\usage{
deletes_duplcated_data(DATA, DATETIME_HEADER = "TIMESTAMP")
}
\arguments{
\item{DATA}{data.frame}

\item{DATETIME_HEADER}{header corresponding to datetime}
}
\value{
A list containing a data.frame without duplicated rows and a dataframe with duplicated rows
}
\description{
This function delete duplcated data in a data.frame
}
\examples{
deletes_duplcated_data(DATA = mydata, DATETIME_HEADER = "TIMESTAMP")
deletes_duplcated_data(DATA = your data.frame, DATETIME_HEADER = "Your datetime header")

}
