% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_out_of_range.R
\name{exclude_out_of_range}
\alias{exclude_out_of_range}
\title{This function reject all values greater than max and smaller than min in a data.frame}
\usage{
exclude_out_of_range(DATA, DATETIME_HEADER = "TIMESTAMP", RANGE_DIR,
  RANGE_FILE)
}
\arguments{
\item{DATA}{data.frame of timeseries}

\item{DATETIME_HEADER}{header of data corresponding to datetime}

\item{RANGE_DIR}{directory where support files are stored}

\item{RANGE_FILE}{name of filr where min/max thresholds are defined for each variable. Thi file is in RANGE_DIR}
}
\value{
A data.frame with all values in its physical range
}
\description{
This function reject all values greater than max and smaller than min in a data.frame
}
\examples{
exclude_out_of_range(DATA = mydata, RANGE_DIR = RANGE_DIR, RANGE_FILE = "Range.csv")

}
