% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_out_of_range.R
\name{exclude_out_of_range}
\alias{exclude_out_of_range}
\title{This function reject all values greater than max and smaller than min in a data.frame}
\usage{
exclude_out_of_range(DATA, SUPPORT_DIR, RANGE_FILE)
}
\arguments{
\item{DATA}{data.frame of timeseries}

\item{SUPPORT_DIR}{directory where support files are stored}

\item{RANGE_FILE}{name of filr where min/max thresholds are defined for each variable. Thi file is in SUPPORT_DIR}
}
\description{
This function reject all values greater than max and smaller than min in a data.frame
}
